// src/interfaces/trending.ts
// Generated by https://quicktype.io

export interface ITrendingList {
    page:          number;
    results:       ITrendingItem[];
    total_pages:   number;
    total_results: number;
}

export interface ITrendingItem {
    adult:             boolean;
    backdrop_path:     string;
    id:                number;
    name?:             string;
    title?:            string;
    original_language: OriginalLanguage;
    original_name?:    string;
    original_title?:   string;
    overview:          string;
    poster_path:       string;
    media_type:        TrendingMediaType;
    genre_ids:         number[];
    popularity:        number;
    first_air_date?:   string;
    release_date?:     string;
    video?:            boolean;
    vote_average:      number;
    vote_count:        number;
    origin_country?:   OriginCountry[];
}

export enum TrendingMediaType {
    Tv = "tv",
    Movie = "movie",
}

export enum OriginCountry {
    GB = "GB",
    Jp = "JP",
    Kr = "KR",
    Us = "US",
}

export enum OriginalLanguage {
    En = "en",
    Ja = "ja",
    Ko = "ko",
    Es = "es",
    Fi = "fi",
    Fr = "fr",
}